+++ src/main/java/no/nav/familie/integrasjoner/kodeverk/CachedKodeverkService.kt
@@ -8,23 +8,18 @@ import org.springframework.cache.annotation.Cacheable
 import org.springframework.stereotype.Service
 
 @Service
-class KodeverkService(val kodeverkClient: KodeverkClient) {
+class CachedKodeverkService(private val kodeverkClient: KodeverkClient) {
 
-    fun hentPoststed(postnummer: String): String {
-        return hentPostnummer().getOrDefault(postnummer, "")
-    }
-
-    @Cacheable("kodeverk_postestedMedHistorikk")
+    @Cacheable("kodeverk_postestedMedHistorikk", sync = true)
     fun hentPostnummerMedHistorikk(): KodeverkDto = kodeverkClient.hentPostnummerMedHistorikk()
 
     @Cacheable("kodeverk_postested")
-    protected fun hentPostnummer(): Map<String, String> = kodeverkClient.hentPostnummer().mapTerm()
-
-    fun hentLandkode(landkode: String): String = hentLandkoder().getOrDefault(landkode, "")
+    fun hentPostnummer(): Map<String, String> = kodeverkClient.hentPostnummer().mapTerm()
 
     @Cacheable("kodeverk_landkoder")
     fun hentLandkoder(): Map<String, String> = kodeverkClient.hentLandkoder().mapTerm()
 
     @Cacheable("kodeverk_landkoderMedHistorikk")
     fun hentLandkoderMedHistorikk(): KodeverkDto = kodeverkClient.hentLandkoderMedHistorikk()
+
 }
